backend/requirements.txt
asgiref==3.11.1
channels==4.3.2
channels_redis==4.3.0
daphne==4.2.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
msgpack==1.1.2
psycopg==3.3.3
psycopg-binary==3.3.3
Pillow==11.2.1
PyJWT==2.12.0
python-dotenv==1.1.1
redis==7.3.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
